web: database: Do not optimize unbound variables away in SPARQL results.
In e02028b4faafc70ea096356966ee0a19f952ed7d we found that not all
output variables in the SPARQL query are present in the resulting
Python dictionary.

This commit implements a mechanism to ensure all output variables from
a SPARQL query have a key in the resulting Python dictionary.

This has a downstream consequence that calls to ‘value_or’ could now
return None where they used to return the default value passed to it
before.  Before applying this change, all cases where None is not
expected must be fixed.

Whether it is a good idea to go ahead with this is up for discussion.

* src/djehuty/web/database.py: Ensure all variables specified in the
  SPARQL query have a key in the resulting Python dictionary.