web: webdav: Implement authentication
Only the depositor or collaborators can access the sotrage via WebDAV.
Currently, there are two limits in this implementation. The first one is
that collaborators can read/write in the WebDAV space regardless their roles.
The second is that the enable link in the dataset edit page should be
clicked in case any of the depositor's or collaborators' API key has
been changed. This is because the access control list mapping dataset's
directory and credentials is updated only when ?enable=1 is used in the
WebDAV URL.

* src/djehuty/web/database.py: Pass 'name' to sessions to search an API
  key having 'webdav' name
* src/djehuty/web/resources/html_templates/depositor/edit-dataset.html:
  New WebDAV tab in the dataset edit page.
* src/djehuty/web/resources/sparql_templates/account_sessions.sparql:
  Add a new FILTER to search 'name' predicate.
* src/djehuty/web/resources/static/js/edit-dataset.js: JS code for the
  new WebDAV tab in the dataset edit page.
* src/djehuty/web/ui.py: Prevent anonymously access to WebDAV.
* src/djehuty/web/wsgi.py: Implement WebDAV authentication.