web: search: Simplify backend search filter processing.
This commit removes `search term parser`, which extracts search filter
conditions from a string of a certain syntax on the backend side.

This commit implements that when the search API is called on the frontend side,
the search filters such as scopes, formats, tags are explicitly included in the
'search_filters' key in JSON in the request payload.

* src/djehuty/web/database.py: Simplify __search_query_to_sparql_filters().
* src/djehuty/web/resources/static/js/search.js: Remove search term
  constructor.
* src/djehuty/web/validator.py: Add search_filters().
* src/djehuty/web/wsgi.py: Remove search term parser.